requirements.txt
flwr
tensorflow-cpu
joblib
scikit-learn
clearml
keras